+++ libid/io/loadfile.cpp
@@ -1403,7 +1403,7 @@ static int find_fractal_info(const std::string &gif_file, //
 
     std::memset(info, 0, sizeof(FractalInfo));
     int fractal_info_len = sizeof(FractalInfo) + (sizeof(FractalInfo) + 254) / 255;
-    std::fseek(s_fp, (long)(-1-fractal_info_len), SEEK_END);
+    std::fseek(s_fp, -1 - fractal_info_len, SEEK_END);
     /* TODO: revise this to read members one at a time so we get natural alignment
        of fields within the FractalInfo structure for the platform */
     file_read(info, 1, sizeof(FractalInfo), s_fp);
@@ -1422,7 +1422,7 @@ static int find_fractal_info(const std::string &gif_file, //
         {
             // allow 512 garbage at eof
             offset += 100; // go back 100 bytes at a time
-            std::fseek(s_fp, (long)(0-offset), SEEK_END);
+            std::fseek(s_fp, 0 - offset, SEEK_END);
             file_read(tmp_buff, 1, 110, s_fp); // read 10 extra for string compare
             for (int i = 0; i < 100; ++i)
             {
@@ -1430,7 +1430,7 @@ static int find_fractal_info(const std::string &gif_file, //
                 {
                     // found header?
                     std::strcpy(info->info_id, INFO_ID);
-                    std::fseek(s_fp, (long)(hdr_offset = i-offset), SEEK_END);
+                    std::fseek(s_fp, hdr_offset = i - offset, SEEK_END);
                     /* TODO: revise this to read members one at a time so we get natural alignment
                         of fields within the FractalInfo structure for the platform */
                     file_read(info, 1, sizeof(FractalInfo), s_fp);
@@ -1452,7 +1452,7 @@ static int find_fractal_info(const std::string &gif_file, //
                  might be over 255 chars, and thus earlier load might be bad
                  find exact endpoint, so scan back to start of ext blks works
                */
-            std::fseek(s_fp, (long)(hdr_offset-15), SEEK_END);
+            std::fseek(s_fp, hdr_offset - 15, SEEK_END);
             int scan_extend = 1;
             while (scan_extend)
             {
@@ -1481,7 +1481,7 @@ static int find_fractal_info(const std::string &gif_file, //
                 case GifExtensionId::RESUME_INFO: // "fractint002", resume info
                     skip_ext_blk(&block_len, &data_len); // once to get lengths
                     blk_2_info->resume_data.resize(data_len);
-                    std::fseek(s_fp, (long)(0-block_len), SEEK_CUR);
+                    std::fseek(s_fp, 0 - block_len, SEEK_CUR);
                     load_ext_blk((char *)g_block, data_len);
                     // resume data is assumed to be platform native; no need to decode
                     std::copy(&g_block[0], &g_block[data_len], blk_2_info->resume_data.data());
@@ -1491,7 +1491,7 @@ static int find_fractal_info(const std::string &gif_file, //
                 case GifExtensionId::FORMULA_INFO: // "fractint003", formula info
                     skip_ext_blk(&block_len, &data_len); // once to get lengths
                     // check data_len for backward compatibility
-                    std::fseek(s_fp, (long)(0-block_len), SEEK_CUR);
+                    std::fseek(s_fp, 0 - block_len, SEEK_CUR);
                     load_ext_blk((char *)&formula_info, data_len);
                     // TODO: decode formula info?
                     std::strcpy(blk_3_info->form_name, formula_info.form_name);
@@ -1524,7 +1524,7 @@ static int find_fractal_info(const std::string &gif_file, //
                     assert(data_len % 2 == 0);  // should specify an integral number of 16-bit ints
                     blk_4_info->length = data_len/2;
                     blk_4_info->range_data.resize(blk_4_info->length);
-                    std::fseek(s_fp, (long) -block_len, SEEK_CUR);
+                    std::fseek(s_fp, -block_len, SEEK_CUR);
                     {
                         std::vector<char> buffer(data_len, 0);
                         load_ext_blk(buffer.data(), data_len);
@@ -1539,14 +1539,14 @@ static int find_fractal_info(const std::string &gif_file, //
                 case GifExtensionId::EXTENDED_PRECISION: // "fractint005", extended precision parameters
                     skip_ext_blk(&block_len, &data_len); // once to get lengths
                     blk_5_info->apm_data.resize(data_len);
-                    std::fseek(s_fp, (long)(0-block_len), SEEK_CUR);
+                    std::fseek(s_fp, 0 - block_len, SEEK_CUR);
                     load_ext_blk(blk_5_info->apm_data.data(), data_len);
                     // TODO: decode extended precision parameters?
                     blk_5_info->got_data = true;
                     break;
                 case GifExtensionId::EVOLVER_INFO: // "fractint006", evolver params
                     skip_ext_blk(&block_len, &data_len); // once to get lengths
-                    std::fseek(s_fp, (long)(0-block_len), SEEK_CUR);
+                    std::fseek(s_fp, 0 - block_len, SEEK_CUR);
                     load_ext_blk((char *)&evolution_info, data_len);
                     decode_evolver_info(&evolution_info, 1);
                     blk_6_info->length = data_len;
@@ -1576,7 +1576,7 @@ static int find_fractal_info(const std::string &gif_file, //
                     break;
                 case GifExtensionId::ORBITS_INFO: // "fractint007", orbits parameters
                     skip_ext_blk(&block_len, &data_len); // once to get lengths
-                    std::fseek(s_fp, (long)(0-block_len), SEEK_CUR);
+                    std::fseek(s_fp, 0 - block_len, SEEK_CUR);
                     load_ext_blk((char *)&orbits_info, data_len);
                     decode_orbits_info(&orbits_info, 1);
                     blk_7_info->length = data_len;
@@ -1654,7 +1654,7 @@ static void skip_ext_blk(int *block_len, int *data_len)
     *block_len = 1;
     while ((len = fgetc(s_fp)) > 0)
     {
-        std::fseek(s_fp, (long)len, SEEK_CUR);
+        std::fseek(s_fp, len, SEEK_CUR);
         *data_len += len;
         *block_len += len + 1;
     }
